# webapp/openapi/integration-v1.yaml
openapi: 3.1.0
...
info:
  title: Kinboard Integration API
  version: "1.0.0"
  description: |
    Machine-to-machine access to a Kinboard household.

    Authentication is a bearer token created in **Settings → Integrations**.
    Tokens carry explicit scopes; no scope implies another, and every failure
    to authorise answers `401` — including "valid token, wrong scope", so a
    caller probing an endpoint cannot learn that its token is real.
  license:
    name: PolyForm Noncommercial 1.0.0
    identifier: PolyForm-Noncommercial-1.0.0
servers:
  - url: "{kinboard_url}/api/integration/v1"
    variables:
      kinboard_url:
        default: "http://kinboard.local:3000"
        description: The address of the Kinboard instance.
security:
  - integrationToken: []
...
paths:
...
  /meals:
    get:
      tags: [state]
      summary: Read planned meals in a date range
      description: |
        Entries overlapping `start`..`end`, both inclusive, both plain
        `YYYY-MM-DD` dates — a meal plan entry has no time-of-day, so
        timestamps are refused rather than silently truncated. Scoped
        through `meal_plans!inner(family_id)`, since `meal_plan_entries`
        carries no `family_id` of its own, and by the same dates the window
        asks for rather than by which `meal_plans` row an entry happens to
        sit in — a household's chosen week start can change at any time.
      x-required-scope: family:read
      parameters:
        - name: start
          in: query
          required: true
          schema: { type: string, format: date }
        - name: end
          in: query
          required: true
          schema: { type: string, format: date }
          description: Must not be before `start`, and the range may not exceed 31 days.
      responses:
        "200":
          description: Entries overlapping the window, earliest date first
          content:
            application/json:
              schema:
                type: object
                required: [entries]
                properties:
                  entries:
                    type: array
                    items: { $ref: "#/components/schemas/MealPlanEntry" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "429": { $ref: "#/components/responses/RateLimited" }
    post:
      tags: [state]
      summary: Add a meal to the plan
      description: |
        Finds or creates the `meal_plans` row for the week containing
        `date`, following the family's `week_start` setting the same way
        `getOrCreateMealPlanId` in `hooks/use-meal-planner.ts` does
        (`locale` resolves to Monday here — there is no device to resolve
        it against on the server).

        Exactly one of `recipe_id` (must name a recipe in this family) or
        `note` (1-200 characters) is required. This adds an entry to the
        slot; it does not replace whatever else is already planned for that
        date and meal type.
      x-required-scope: meals:write
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, minLength: 8, maxLength: 255 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [date, meal_type]
              properties:
                date: { type: string, format: date }
                meal_type: { type: string, enum: [breakfast, lunch, dinner, snack] }
                recipe_id: { type: string, format: uuid }
                note: { type: string, minLength: 1, maxLength: 200 }
                servings: { type: integer, minimum: 1, maximum: 50 }
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                type: object
                required: [entry]
                properties:
                  entry: { $ref: "#/components/schemas/MealPlanEntry" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { description: recipe_id does not name a recipe in this family. }
        "409": { description: Idempotency key reused with different arguments. }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { description: The entry could not be stored. }
...
  /cameras:
    get:
      tags: [state]
      summary: The family's cameras, as show_camera takes them
      description: |
        Every enabled camera from Settings → Cameras, in the order the
        settings page shows them: its id and its name, and nothing else. A
        camera's stream URL names the household's network and often carries
        its credentials, so it never leaves the server.

        `show_camera` takes either the `id` or the exact `name`.
      x-required-scope: family:read
      responses:
        "200":
          description: The family's cameras
          content:
            application/json:
              schema:
                type: object
                required: [cameras]
                properties:
                  cameras:
                    type: array
                    items: { $ref: "#/components/schemas/CameraRef" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /birthdays:
    get:
      tags: [state]
      summary: List the family's birthdays
      description: |
        Every birthday not in the recycle bin, sorted by `days_until` (the
        next one first; ties by name). `today` is the family's date in its
        time zone, which `next_date`, `days_until`, `age` and `turns` are
        computed from. 29 February falls on 1 March in a year without one,
        as everywhere else in Kinboard.

        Kinboard has no separate "unknown year" field: the birthdays page
        stores a blank year as the current year, and a stored year that is
        not before the current one reads as unknown. Such a birthday comes
        back as `--MM-DD` with `year_known: false` and `age`/`turns` null.
        A child born this year reads the same way.
      x-required-scope: family:read
      responses:
        "200":
          description: The family's birthdays
          content:
            application/json:
              schema:
                type: object
                required: [today, birthdays]
                properties:
                  today: { type: string, format: date }
                  birthdays:
                    type: array
                    items: { $ref: "#/components/schemas/Birthday" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { description: The birthdays could not be read. }
    post:
      tags: [state]
      summary: Add a birthday
      description: |
        Adds a birthday as the birthdays page does. `date` is `YYYY-MM-DD`
        with a birth year from 1900 to last year (in the family's time
        zone), or `--MM-DD` when the year is unknown — stored with the
        current year, as the page stores a blank year, so `--02-29` is
        refused in a year that is not a leap year. A full date with this
        year's birth year is refused (`400`, pointing at `--MM-DD`): it
        would be stored exactly like an unknown year and read back without
        it. A later year is refused as a date in the future. `person_id` must name a person of this
        family who is not in the recycle bin. `notify_days_before` defaults
        to 7. Nothing is written when any field is refused.
      x-required-scope: birthdays:write
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, minLength: 8, maxLength: 255 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, date]
              properties:
                name: { type: string, minLength: 1, maxLength: 100, description: Trimmed first. }
                date: { $ref: "#/components/schemas/BirthdayDate" }
                person_id: { type: [string, "null"], format: uuid }
                notify_days_before: { type: integer, minimum: 0, maximum: 60, default: 7 }
      responses:
        "201":
          description: The birthday was added
          content:
            application/json:
              schema:
                type: object
                required: [birthday]
                properties:
                  birthday: { $ref: "#/components/schemas/Birthday" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "409": { description: Idempotency key reused with different arguments. }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { description: The birthday could not be added. }
...
components:
  securitySchemes:
    integrationToken:
      type: http
      scheme: bearer
      description: A token from Settings → Integrations. Shown once; stored only as a hash.
  responses:
    Unauthenticated:
      description: |
        No token, an unknown/revoked/expired token, or a token without the
        required scope. Deliberately one answer for all four — distinguishing
        them would tell a caller probing an endpoint that its token is real.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    InvalidRequest:
      description: The request was understood but its arguments were not usable.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: |
        Too many requests for this token. Budgets are per token and per method
        class: reads and writes are counted separately, so an automation that
        writes too often does not also blind a dashboard that only reads.

        `Retry-After` gives the seconds to wait and is never zero.
      headers:
        Retry-After:
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    MealPlanEntry:
      type: object
      required: [id, date, meal_type]
      properties:
        id: { type: string, format: uuid }
        date: { type: string, format: date }
        meal_type: { type: string, enum: [breakfast, lunch, dinner, snack] }
        recipe_id: { type: [string, "null"], format: uuid }
        recipe_title:
          type: [string, "null"]
          description: Null when the entry has no linked recipe.
        note: { type: [string, "null"] }
        servings: { type: [integer, "null"] }
...
    CameraRef:
      type: object
      description: A camera as the Integration API knows it. Never its stream URL.
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable. May be reworded at any time — do not branch on it.
        code:
          type: string
          description: Stable and machine-readable. Branch on this.
          enum:
            - not_authenticated
            - forbidden
            - not_found
            - invalid_request
            - conflict
            - rate_limited
            - upstream_unavailable
            - unavailable
            - internal_error
            - too_many_timers
            - not_implemented
        correlationId:
          type: string
          description: Also sent as the `x-correlation-id` header, and printed on every log line for this request.
...
    BirthdayDate:
      type: string
      pattern: "^(\\d{4}|-)-\\d{2}-\\d{2}$"
      description: "`YYYY-MM-DD`, or `--MM-DD` when the birth year is unknown."
    Birthday:
      type: object
      required: [id, name, date, year_known, next_date, days_until, age, turns, person_id, notify_days_before]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        date: { $ref: "#/components/schemas/BirthdayDate" }
        year_known: { type: boolean }
        next_date: { type: string, format: date, description: "The day it next falls on; today counts." }
        days_until: { type: integer, minimum: 0 }
        age: { type: [integer, "null"], description: Age today; null without a birth year. }
        turns: { type: [integer, "null"], description: "Age on `next_date`; null without a birth year." }
        person_id: { type: [string, "null"], format: uuid }
        notify_days_before: { type: integer, minimum: 0 }
...